fix: prevent server restarts in watch mode
If rollup is in watch mode and the rollup.config is changed, this
triggers a restart of the livereload server.  This restart can reset
the connection between the browser and the server requiring users to
manually reload.

Now the server is cached in memory and subsequent livereload() calls
use that version.